+++ common/src/main/kotlin/net/stellarica/common/coord/Position.kt
@@ -1,0 +1,4 @@
+package net.stellarica.common.coord
+
+class Position {
+}
