+++ server/src/main/kotlin/net/stellarica/server/persistence/ChunkPersistentStorage.kt
@@ -1,4 +1,4 @@
-package net.stellarica.server.util.persistence;
+package net.stellarica.server.persistence;
 
 import org.bukkit.Chunk
 import org.bukkit.persistence.PersistentDataContainer
