+++ server/src/main/kotlin/net/stellarica/server/persistence/WorldPersistentStorage.kt
@@ -1,4 +1,4 @@
-package net.stellarica.server.util.persistence
+package net.stellarica.server.persistence
 
 import org.bukkit.World
 import org.bukkit.persistence.PersistentDataContainer
