+++ server/src/main/kotlin/net/stellarica/server/util/World.kt
@@ -1,0 +1,20 @@
+package net.stellarica.server.util
+
+import net.stellarica.server.material.type.block.BlockType
+import org.bukkit.block.BlockState
+
+class World {
+    fun getBlockAt(pos: BlockPosition): BlockType {
+
+    }
+
+    fun getBlockStateAt(pos: BlockPosition): BlockState {
+
+    }
+
+    fun setBlockAt(pos: BlockPosition, type: BlockType) {
+
+    }
+
+    fun moveBlock(from: BlockPosition, to: BlockPosition) {}
+}
